+++ components/Map.web.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+export default function Map() {
+  return (
+    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+      <Text>Map View is not available on the web.</Text>
+    </View>
+  );
+}
